requirements.txt
# Core de TensorFlow
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.37.1

# Ecosistema de datos/tf
keras==2.10.0
Keras-Preprocessing==1.1.2
tensorflow-datasets==4.8.2         # requiere protobuf>=3.12.2                     :contentReference[oaicite:0]{index=0}
tensorflow-metadata==1.12.0        # requiere protobuf>=3.13,<4                   :contentReference[oaicite:1]{index=1}

# Parsiado y serialización (Protobuf)
protobuf==3.19.6                   # cumple <3.20 para TF y ≥3.13 para TF Metadata

# Abseil (TF Metadata exige <2.0.0)
absl-py==1.4.0                     # <2.0.0 para ser compatible con TF Metadata

# Numpy y SciPy
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.6.1
scipy==1.13.1
h5py==3.14.0

# Utilidades de plotting
matplotlib==3.9.4
Pillow==9.3.0
kiwisolver==1.4.7
cycler==0.12.1
contourpy==1.3.0
fonttools==4.58.5

# Lógica y DSLs
astunparse==1.6.3
gast==0.4.0

# Concurrencia y asincronía
grpcio==1.73.1
threadpoolctl==3.6.0

# Redes y autenticación
requests==2.32.4
urllib3==2.5.0
certifi==2025.7.9
charset-normalizer==3.4.2
idna==3.10
oauthlib==3.3.1
requests-oauthlib==2.0.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0

# Herramientas varias
cachetools==5.5.2
click==8.1.8
dill==0.3.6                 # instalado automáticamente por tf-datasets
etils==1.5.2
flatbuffers==25.2.10
fsspec==2025.5.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.1
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
packaging==25.0
promise==2.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rsa==4.9.1
six==1.17.0
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
